add toas

+++ src/ft_wcslen.c
@@ -6,20 +6,18 @@
 /*   By: chle-van <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/30 13:28:51 by chle-van          #+#    #+#             */
-/*   Updated: 2017/03/30 13:46:06 by chle-van         ###   ########.fr       */
+/*   Updated: 2017/05/27 23:06:29 by chle-van         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_wcslen(wchar_t *wcs)
+size_t	ft_wcslen(wchar_t *wcs)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	while (*wcs && wcs[i])
 		i++;
 	return (i);
-		
-
 }
